Safari-609.1.11.1/LayoutTests/fast/events/touch/ios/touch-event-regions/scrolled-overflow.html

<!DOCTYPE html>
<html>
<head>
    <meta name="viewport" content="initial-scale=1.0">
    <style>
        #scroller {
            height: 200px;
            width: 200px;
            border: 1px solid black;
            margin: 50px;
            overflow: scroll;
        }
        .box {
            position: relative;
            height: 100px;
            width: 120px;
            border: 1px solid black;
            margin: 250px 10px;
        }
    </style>
    <script src="resources/touch-regions-helper.js"></script>
    <script>
        if (window.testRunner)
            testRunner.dumpAsText();

        window.addEventListener('load', () => {
            scroller.scrollTo(0, 200);
            dumpRegions();
        }, false);
    </script>
</head>
<body>
    <div id="scroller">
        <div class="box" ontouchstart="(void)0">touchstart inside overflow</div>
    </div>
</body>
</html>
