Safari-609.1.11.1/LayoutTests/fast/events/touch/ios/touch-event-regions/scrolled-overflow-expected.txt

touchstart inside overflow
touchstart
58, 50 - 260, 252
touchstart
58, 50 - 260, 252
touchend
58, 50 - 260, 252
touchforcechange
58, 50 - 260, 252
passive
